requirements.txt
BeautifulSoup==3.2.1
Django==1.6.1
Fabric==1.8.1
MySQL-python==1.2.5
Pillow==2.3.0
Pygments==1.6
South==0.8.4
amqp==1.4.1
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.14
bpython==0.12
celery==3.1.8
distribute==0.7.3
django-celery==3.1.1
django-extensions==1.3.2
django-globals==0.1.0
django-markitup==2.1
django-robots==1.0
docutils==0.11
ecdsa==0.10
fish==1.1
gunicorn==18.0
ipython==1.1.0
kombu==3.0.10
micawber==0.3.0
paramiko==1.12.1
pycrypto==2.6.1
pylibmc==1.2.3
python-dateutil==2.2
pytz==2013.9
six==1.5.2
wsgiref==0.1.2
